Changes to run regTests on Jenkins
1. Shell script & bitcoin.conf file from OmniJ
2. Rename/update shell script
3. Gradle fixes 
4. Run tests from two modules cli and rpcclient

+++ bitcoinj-rpcclient/run-bitcoind-regtest.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+#
+# This script is intended to run on Jenkins server to test `bitcoind` in RegTest mode
+# using the bitcoinj-rpcclient.
+#
+# WARNING: Read this script carefully before running or using. Note the `rm -rf` command.
+#
 set -x
 
 function cleanup {
@@ -9,19 +15,25 @@ trap cleanup EXIT
 BTCD=copied-artifacts/src/bitcoind
 DATADIR=$HOME/.bitcoin
 LOGDIR=logs
-MSCLOG=/tmp/mastercore.log
+#MSCLOG=/tmp/mastercore.log
+
+if [ ! -f "$BTCD" ]
+then
+  echo "$BTCD doesn't exist - check your Jenkins test setup."
+  exit 1
+fi
 
 # Assume bitcoind built elsewhere and copied by Jenkins Copy Artifact plugin
 chmod +x $BTCD
 
 # Setup bitcoin conf and data dir
 mkdir -p $DATADIR
-cp -n bitcoin.conf $DATADIR
+cp -n bitcoinj-rpcclient/bitcoin.conf $DATADIR
 
 # Setup logging
 mkdir -p $LOGDIR
-touch $MSCLOG
-ln -sf $MSCLOG $LOGDIR/mastercore.log
+#touch $MSCLOG
+#ln -sf $MSCLOG $LOGDIR/mastercore.log
 
 # Remove all regtest data
 rm -rf $DATADIR/regtest
@@ -31,12 +43,9 @@ $BTCD -regtest -datadir=$DATADIR > $LOGDIR/bitcoin.log &
 BTCSTATUS=$?
 BTCPID=$!
 
-# Give server some time to start
-#sleep 30
-
 # Run integration tests
 echo "Running Bitcoin RPC integration tests in regtest mode..."
-./gradlew clean :bitcoin-rpc:regTest
+./gradlew clean :bitcoin-rpcclient:regTest :bitcoinj-cli:regTest
 GRADLESTATUS=$?
 
 exit $GRADLESTATUS
